Machine Learning Code
Implemented Random Forest, SVM as well as a stacking classifier which combines the two using sk-learn. I entered the competition with name world_Hello and was ranked 16th out of the whole class. Refer to the following link for further details about the competition and the dataset used: https://www.kaggle.com/competitions/i-wish-i-was-there/leaderboard